Change permissions field from Int to Int64

+++ src/types/guild.jl
@@ -64,7 +64,7 @@ struct Guild <: AbstractGuild
     splash::OptionalNullable{String}
     owner::Optional{Bool}
     owner_id::Optional{Snowflake}  # Missing in Invite.
-    permissions::Optional{Int}
+    permissions::Optional{Int64}
     region::Optional{String}  # Invite
     afk_channel_id::OptionalNullable{Snowflake}  # Invite
     afk_timeout::Optional{Int}  # Invite
